Safari-611.1.10/LayoutTests/fast/dom/location-valueOf-after-object-prototype-update2.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests the value of location.valueOf after Object.prototype.valueOf has been replaced.");

var objectProtoValueOf = Object.prototype.valueOf;
delete Object.prototype.valueOf;
Object.defineProperty(Object.prototype, "valueOf", {value: 3});
shouldBe("location.valueOf", "objectProtoValueOf");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
